[feature] IPeerView
