+++ core/network/peer_view.hpp
@@ -5,22 +5,12 @@
  */
 #pragma once
 
-#include <memory>
-#include <unordered_map>
-
-#include <libp2p/peer/peer_id.hpp>
-
-#include "application/app_state_manager.hpp"
-#include "blockchain/block_tree.hpp"
-#include "crypto/type_hasher.hpp"
-#include "injector/lazy.hpp"
-#include "network/types/collator_messages.hpp"
-#include "outcome/outcome.hpp"
-#include "primitives/event_types.hpp"
-#include "subscription/subscriber.hpp"
-#include "subscription/subscription_engine.hpp"
+#include "network/i_peer_view.hpp"
 #include "utils/non_copyable.hpp"
 #include "utils/safe_object.hpp"
+#include "injector/lazy.hpp"
+#include "application/app_state_manager.hpp"
+#include "blockchain/block_tree.hpp"
 
 namespace kagome::blockchain {
   class BlockTree;
@@ -28,43 +18,14 @@ namespace kagome::blockchain {
 
 namespace kagome::network {
 
-  using HashedBlockHeader = primitives::BlockHeader;
-  struct ExView {
-    View view;
-    HashedBlockHeader new_head;
-    std::vector<primitives::BlockHash> lost;
-  };
-
-  struct ExViewRef {
-    std::optional<std::reference_wrapper<const HashedBlockHeader>> new_head;
-    const std::vector<primitives::BlockHash> &lost;
-  };
-
   /**
    * Observable class for current heads and finalized block number tracking.
    */
   class PeerView final : public NonCopyable,
                          public NonMovable,
+                         public IPeerView,
                          public std::enable_shared_from_this<PeerView> {
    public:
-    enum struct EventType : uint32_t { kViewUpdated, kPeerRemoved };
-
-    using PeerId = libp2p::peer::PeerId;
-
-    using MyViewSubscriptionEngine = subscription::
-        SubscriptionEngine<EventType, std::monostate, network::ExView>;
-    using MyViewSubscriptionEnginePtr =
-        std::shared_ptr<MyViewSubscriptionEngine>;
-    using MyViewSubscriber = MyViewSubscriptionEngine::SubscriberType;
-    using MyViewSubscriberPtr = std::shared_ptr<MyViewSubscriber>;
-
-    using PeerViewSubscriptionEngine = subscription::
-        SubscriptionEngine<EventType, std::monostate, PeerId, network::View>;
-    using PeerViewSubscriptionEnginePtr =
-        std::shared_ptr<PeerViewSubscriptionEngine>;
-    using PeerViewSubscriber = PeerViewSubscriptionEngine::SubscriberType;
-    using PeerViewSubscriberPtr = std::shared_ptr<PeerViewSubscriber>;
-
     PeerView(primitives::events::ChainSubscriptionEnginePtr chain_events_engine,
              std::shared_ptr<application::AppStateManager> app_state_manager,
              LazySPtr<blockchain::BlockTree> block_tree);
@@ -76,14 +37,13 @@ namespace kagome::network {
     bool prepare();
     void stop();
 
-    size_t peersCount() const;
+    size_t peersCount() const override;
+    MyViewSubscriptionEnginePtr getMyViewObservable() override;
+    PeerViewSubscriptionEnginePtr getRemoteViewObservable() override;
 
-    MyViewSubscriptionEnginePtr getMyViewObservable();
-    PeerViewSubscriptionEnginePtr getRemoteViewObservable();
-
-    void removePeer(const PeerId &peer_id);
-    void updateRemoteView(const PeerId &peer_id, network::View &&view);
-    auto &getMyView() const {
+    void removePeer(const PeerId &peer_id) override;
+    void updateRemoteView(const PeerId &peer_id, network::View &&view) override;
+    const View &getMyView() const override {
       return my_view_;
     }
